requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
ifaddr==0.1.7
ipaddress==1.0.23
lxml==4.9.4
miniupnpc==2.3.2
packaging==24.2
pip-review==1.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
six==1.17.0
upnpclient==1.0.3
urllib3==2.4.0
